git add . git commit -m "plz work"
